--HG--
branch : yt

+++ yt/data_objects/object_finding_mixin.py
@@ -26,7 +26,7 @@ License:
 import numpy as na
 
 from yt.funcs import *
-from yt.utilities.amr_utils import \
+from yt.utilities.lib import \
     get_box_grids_level, \
     get_box_grids_below_level
 
@@ -212,7 +212,7 @@ class ObjectFindingMixin(object):
         get_box_grids_below_level(left_edge, right_edge,
                             level,
                             self.grid_left_edge, self.grid_right_edge,
-                            self.grid_levels, mask, min_level)
+                            self.grid_levels.astype("int32"), mask, min_level)
         mask = mask.astype("bool")
         return self.grids[mask], na.where(mask)
